+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1 Down the Drain
+
+- Queue is an emitter
+- Queue emits `drain` events
+
 # 1.1.0 Obama Cares
 
 **BREAKING**